Rework skip_card, introduce ParsedLine
NoCommenIter now returns ParsedLine, i.e. all lines are parsed, which
pessimizes skip_ges performance, which basically doesn't influcence
anything at all.

skip_ges was reworked to take a ParsedLine argument and deal with all
the gnarly details itself, and returns an Option<SkipResult> for the
cases where the ParsedLine is garbage and the NoCommentIterator is not
advanced.

Introduce several macros to improve code structure(?).