Read buffer data from the file directly
It's faster in the benchmarks, as we can basically be sure the file is
on the OS cache anyways. Need further confirmation for different
environments, though.

Technically, we've introduced `Line`, which can either hold a `&[u8]` or
a `String`. Reading from the file gives us a `Vec<u8>` (so we can skip
UTF-8 validation), which we split into the initial `Vec<Line>` where
each entry is a `&[u8]` under the hood. Updating from a `Vec<String>`
via the buffer updates API inserts the `String` variant into our `Lines`
vec.